+++ src/message.java
@@ -3,7 +3,7 @@ import java.math.BigInteger;
 public class message {
     String message;
     Account account;
-    int realNumber;
+    long realNumber;
     public message (String message, Account account, int realNumber){
         this.message = message;
         this.account = account;
